Upgrade to latest core workings
A new Vagrant box has been setup on Hashicorp's Atlas (Vreasy/vrrecruit).
The stack differs a bit from ours, but it is enough for the assigment
to be completed.

All our core classes have been updated, including the Persistence trait
and the Rest-like controller.

The old angular structure has been left as it was, so to avoid complicating
things further when adding coffeescript.